+++ src/lib/File.h
@@ -10,6 +10,6 @@ class File
 public:
     static bool Exists(const wchar_t* filePath);
     static void ErrorIfFileNotFound(const wchar_t* filePath);
-    static int GetFileSizeBytes(const wchar_t* filePath);
+    static long long GetFileSizeBytes(const wchar_t* filePath);
     static int ReadAllBytes(const wchar_t* filePath, int bufferSize, void* buffer);
 };
